AIFan/backend/src/blockchain/raydium_client.py
from solana.rpc.api import Client
from solana.transaction import Transaction
import os
from dotenv import load_dotenv

class RaydiumClient:
    def __init__(self):
        load_dotenv()
        self.client = Client(os.getenv("RPC_URL"))
        
    def create_liquidity_pool(self, token_mint, sol_amount, token_amount):
        """Create a liquidity pool on Raydium"""
        try:
            # Implementation for Raydium pool creation
            # This will require interaction with Raydium's smart contracts
            pass
        except Exception as e:
            return {
                "success": False,
                "error": str(e)
            }
            
    def add_liquidity(self, pool_id, sol_amount, token_amount):
        """Add liquidity to existing pool"""
        try:
            # Implementation for adding liquidity
            pass
        except Exception as e:
            return {
                "success": False,
                "error": str(e)
            }
        
    def get_pool_info(self, pool_id):
        """Get pool information"""
        try:
            # Implementation for getting pool info
            pass
        except Exception as e:
            return {
                "success": False,
                "error": str(e)
            }

    def calculate_price_impact(self, token_amount, sol_amount):
        """Calculate price impact for liquidity addition"""
        try:
            # Calculate price impact before pool creation
            pass
        except Exception as e:
            return {
                "success": False,
                "error": str(e)
            }