AIFan/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1

sqlalchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9
supabase==2.3.4

openai==1.12.0
elevenlabs==1.13.1
transformers==4.37.2

pydantic>=2.5.3,<3.0.0
pydantic-core>=2.18.2,<3.0.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
requests==2.31.0
tweepy==4.14.0
